# lib/api-zod/src/generated/types/classificationPreview.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Directory Master API
 * OpenAPI spec version: 0.1.0
 */
import type { LocationSuggestion } from "./locationSuggestion";
import type { ServiceSuggestion } from "./serviceSuggestion";

export interface ClassificationPreview {
  entryId: number;
  entryTitle: string;
  locationSuggestion?: LocationSuggestion | null;
  serviceSuggestions: ServiceSuggestion[];
}