# lib/api-zod/src/generated/types/entryClassificationResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Directory Master API
 * OpenAPI spec version: 0.1.0
 */
import type { EntryClassificationResultServicesItem } from "./entryClassificationResultServicesItem";
import type { NormalizedLocation } from "./normalizedLocation";

export interface EntryClassificationResult {
  entryId: number;
  location?: NormalizedLocation | null;
  services: EntryClassificationResultServicesItem[];
}