# lib/api-zod/src/generated/types/localSeoApplyResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Directory Master API
 * OpenAPI spec version: 0.1.0
 */

export interface LocalSeoApplyResponse {
  success: boolean;
  locationsApplied: number;
  servicesQueued: number;
}