# lib/api-zod/src/generated/types/localSeoCoverageResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Directory Master API
 * OpenAPI spec version: 0.1.0
 */

export interface LocalSeoCoverageResponse {
  totalEntries: number;
  confirmedLocations: number;
  locationReviewQueue: number;
  confirmedServices: number;
  serviceReviewQueue: number;
  eligibleStateHubs: number;
  eligibleCityHubs: number;
}