# lib/api-zod/src/generated/types/localSeoPreviewInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Directory Master API
 * OpenAPI spec version: 0.1.0
 */

export interface LocalSeoPreviewInput {
  /** @nullable */
  limit?: number | null;
}