# lib/api-zod/src/generated/types/localSeoPreviewResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Directory Master API
 * OpenAPI spec version: 0.1.0
 */
import type { ClassificationPreview } from "./classificationPreview";

export interface LocalSeoPreviewResponse {
  previews: ClassificationPreview[];
}