# lib/api-zod/src/generated/types/localSeoReviewResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Directory Master API
 * OpenAPI spec version: 0.1.0
 */
import type { LocalSeoReviewRow } from "./localSeoReviewRow";

export interface LocalSeoReviewResponse {
  rows: LocalSeoReviewRow[];
  total: number;
  page: number;
  totalPages: number;
}