# lib/api-zod/src/generated/types/localSeoReviewRow.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Directory Master API
 * OpenAPI spec version: 0.1.0
 */
import type { LocalSeoReviewRowServiceSuggestionsItem } from "./localSeoReviewRowServiceSuggestionsItem";
import type { NormalizedLocation } from "./normalizedLocation";

export interface LocalSeoReviewRow {
  entryId: number;
  /** @nullable */
  entryTitle?: string | null;
  /** @nullable */
  originalLocation?: string | null;
  location?: NormalizedLocation | null;
  serviceSuggestions: LocalSeoReviewRowServiceSuggestionsItem[];
}